requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
backports-abc==0.4        # via tornado
backports.ssl-match-hostname==3.5.0.1  # via tornado, websocket-client
certifi==2016.2.28        # via tornado
contextlib2==0.5.1        # via vcrpy
funcsigs==0.4             # via mock
mock==1.3.0               # via vcrpy
mopidy==2.0.0
pbr==1.8.1                # via mock
py==1.4.31                # via pytest
Pykka==1.2.1              # via mopidy
pytest==2.9.0
PyYAML==3.11              # via vcrpy
requests==2.9.1           # via mopidy, slackclient
singledispatch==3.4.0.3   # via tornado
six==1.10.0               # via mock, singledispatch, vcrpy, websocket-client
slackclient==1.0.0
tornado==4.3              # via mopidy
vcrpy==1.7.4
websocket-client==0.35.0  # via slackclient
wrapt==1.10.6             # via vcrpy

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via mopidy
